feat: Extract payment status and paid_at change logic into custom hooks
- Moved the payment status and paid_at change handling logic from the ApplicationsTable component into new custom hooks, usePaymentStatusChange and usePaidAtChange, for improved separation of concerns and code organization.
- These hooks include validation and state management for updating payment status and paid_at date, enhancing maintainability and clarity in the ApplicationsTable component.
